pyproject.toml
[project]
name = "th8-dcc-backend"
version = "0.3.0"
description = "TH8 Sense DCC Backend (Ingestion + Evidence)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "pydantic>=2.6",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "supabase>=2.6.0",
  "httpx>=0.27",
  "langchain>=0.2.14",
  "langgraph>=0.2.35",
  "langchain-openai>=0.1.22",
  "llama-index-core>=0.11.0",
  "llama-parse>=0.6.0",
]
